Move type so ESM works better.
