--- packages/core/module.js
@@ -1,18 +1,0 @@
-/**
- * @copyright   2017-2019, Miles Johnson
- * @license     https://opensource.org/licenses/MIT
- */
-
-// Our index re-exports TypeScript types, which Babel is unable to detect and omit.
-// Because of this, Webpack and other bundlers attempt to import values that do not exist.
-// To mitigate this issue, we need this module specific index file that manually exports.
-
-import Aesthetic from './esm/Aesthetic';
-import ClassNameAesthetic from './esm/ClassNameAesthetic';
-import UnifiedSyntax from './esm/UnifiedSyntax';
-import Ruleset from './esm/Ruleset';
-import Sheet from './esm/Sheet';
-
-export { ClassNameAesthetic, UnifiedSyntax, Ruleset, Sheet };
-
-export default Aesthetic;
