+++ packages/core/package.json
@@ -20,7 +20,7 @@
   "repository": "https://github.com/milesj/aesthetic/tree/master/packages/core",
   "license": "MIT",
   "main": "./lib/index.js",
-  "module": "./module.js",
+  "module": "./esm/index.js",
   "types": "./lib/index.d.ts",
   "sideEffects": false,
   "publishConfig": {
