+++ packages/core/src/Aesthetic.tsx
@@ -5,6 +5,7 @@ import Sheet from './Sheet';
 import StyleSheetManager from './StyleSheetManager';
 import UnifiedSyntax from './UnifiedSyntax';
 import {
+  AestheticOptions,
   ClassName,
   GlobalSheetDefinition,
   TransformOptions,
@@ -14,17 +15,6 @@ import {
   StyleSheetDefinition,
   ThemeName,
 } from './types';
-
-export interface AestheticOptions {
-  cxPropName: string;
-  extendable: boolean;
-  passThemeProp: boolean;
-  pure: boolean;
-  rtl: boolean;
-  stylesPropName: string;
-  theme: ThemeName;
-  themePropName: string;
-}
 
 export default abstract class Aesthetic<
   Theme extends object,
